+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Unreleased
 - Add `ImpureResponseDecoderInterface`
+- Show detailed type for `DecodedData` value object
 
 ## 2.2.0 - 2022-03-31
 - Use internal cache for `CacheKey` hashed value